FIX error: casting to the same type is unnecessary (`f64` -> `f64`)   --> aws-smithy-types/src/number.rs:41:33    | 41 |             Number::Float(v) => v as f64,    |                                 ^^^^^^^^ help: try: `v`    |    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast    = note: `-D clippy::unnecessary-cast` implied by `-D warnings`

+++ rust-runtime/aws-smithy-types/src/number.rs
@@ -38,7 +38,7 @@ impl Number {
         match self {
             Number::PosInt(v) => v as f64,
             Number::NegInt(v) => v as f64,
-            Number::Float(v) => v as f64,
+            Number::Float(v) => v,
         }
     }
 
@@ -48,7 +48,7 @@ impl Number {
         match self {
             Number::PosInt(v) => v as f32,
             Number::NegInt(v) => v as f32,
-            Number::Float(v) => v as f32,
+            Number::Float(v) => v,
         }
     }
 }
